Reformatted some code, added some JsDoc. Moved operations on local storage to a specific class.
