+++ frontend/src/scripts/DAO/User.js
@@ -1,8 +1,9 @@
 import {api} from './API.js'
+import Formatters from "../Utility/Formatters.js";
 
 class User {
 
-    static BASE_PATH = "/users"
+    static BASE_PATH = "/users";
 
     static loginUser(login, password) {
         return api.request(
@@ -10,7 +11,7 @@ class User {
             api.AUTH_BASE_PATH + '/login',
             `login=${login}&password=${password}`,
             api.CONTENT_TYPE.URL_ENCODED
-        )
+        );
     }
 
     static logoutUser() {
@@ -18,83 +19,75 @@ class User {
             api.METHODS.POST,
             api.AUTH_BASE_PATH + '/logout',
             JSON.stringify({
-                access_token: User.getAccessToken(),
-                refresh_token: User.getRefreshToken()
+                access_token: LocalStorage.getAccessToken(),
+                refresh_token: LocalStorage.getRefreshToken()
             }),
             api.CONTENT_TYPE.JSON
-        )
-    }
-
-    static saveToLocalStorage(access_token, refresh_token){
-        window.localStorage.setItem("access_token", access_token)
-        window.localStorage.setItem("refresh_token", refresh_token)
-    }
-
-    static cleanLocalStorage(){
-        window.localStorage.clear()
+        );
     }
 
     static async connectUser(login, password){
         // We use this function to replace eventual '+' signs in the password
         // They have a special encoding in the URLFormEncoded which is '%2B'
-        password = User.convertPlusToEncoded(password)
+        password = Formatters.convertPlusToEncoded(password);
 
         // Trying to log in user
         let promise = User.loginUser(login, password)
             .then((data) => {
                 if (data.error_code === undefined) {
-                    User.saveToLocalStorage(data.access_token, data.refresh_token)
+                    LocalStorage.saveToLocalStorage(data.access_token, data.refresh_token);
                 }
-                return data
-            })
-        return await promise
+                return data;
+            });
+        return await promise;
     }
 
     static async logout() {
-        await User.logoutUser()
-        User.cleanLocalStorage()
-        window.location.reload()
+        await User.logoutUser();
+        LocalStorage.cleanLocalStorage();
+        window.location.reload();
     }
 
     static isUserConnected() {
-        return (User.getAccessToken() !== null && User.getRefreshToken() !== null)
+        return (LocalStorage.getAccessToken() !== null && LocalStorage.getRefreshToken() !== null);
     }
 
-    static getAccessToken() {
-        return window.localStorage.getItem("access_token")
-    }
 
-    static getRefreshToken() {
-        return window.localStorage.getItem("refresh_token")
-    }
 
     static registerUser(user) {
-        const body = JSON.stringify(user)
-        return api.request(api.METHODS.POST, "/users/register", body)
+        const body = JSON.stringify(user);
+        return api.request(
+            api.METHODS.POST,
+            "/users/register",
+            body
+        );
     }
 
     static fetchUserData() {
-        return api.request_logged(api.METHODS.GET, "/users/me")
+        return api.request_logged(
+            api.METHODS.GET,
+            "/users/me"
+        );
     }
 
     static async refreshTokens() {
         // Trying to log in user
         return api.request(api.METHODS.POST,
             api.AUTH_BASE_PATH + "/refresh",
-            JSON.stringify({"refresh_token": User.getRefreshToken()})
+            JSON.stringify({"refresh_token": LocalStorage.getRefreshToken()})
         ).then((response) => {
             // If the refresh was unsuccessful, we disconnect the user.
             if (response.error_code !== undefined && response.error_code === 401) {
-                User.cleanLocalStorage()
-                window.location.reload()
+                LocalStorage.cleanLocalStorage();
+                window.location.reload();
             } else {
-                User.cleanLocalStorage()
-                User.saveToLocalStorage(
+                LocalStorage.cleanLocalStorage()
+                LocalStorage.saveToLocalStorage(
                     response['access_token'],
                     response['refresh_token']
                 )
             }
-        })
+        });
     }
 
     static async modifyUser() {
@@ -102,11 +95,11 @@ class User {
             api.METHODS.PATCH,
             '/users/',
             JSON.stringify({
-                access_token: User.getAccessToken(),
-                refresh_token: User.getRefreshToken()
+                access_token: LocalStorage.getAccessToken(),
+                refresh_token: LocalStorage.getRefreshToken()
             }),
             api.CONTENT_TYPE.JSON
-        )
+        );
     }
 }
